+++ packages/geoview-core/src/core/utils/config/config-validation.ts
@@ -420,7 +420,7 @@ export class ConfigValidation {
       const propagateLocalizedString = (config: TypeJsonObject): void => {
         if (typeof config === 'object') {
           Object.keys(config).forEach((key) => {
-            if (!key.startsWith('_') && config[key] !== null  && typeof config[key] === 'object') {
+            if (!key.startsWith('_') && config[key] !== null && typeof config[key] === 'object') {
               // Leaving the commented line here in case a developer needs to quickly uncomment it again to troubleshoot
               // logger.logDebug(`Key=${key}`, config[key]);
               if (config?.[key]?.en || config?.[key]?.fr)
