+++ packages/geoview-core/src/geo/layer/gv-layers/raster/gv-wms.ts
@@ -169,7 +169,7 @@ export class GVWMS extends AbstractGVRaster {
           }
         } else featureMember = { plain_text: { '#text': response.data } };
         if (featureMember) {
-          const featureInfoResult = this.#formatWmsFeatureInfoResult(featureMember, layerConfig, clickCoordinate);
+          const featureInfoResult = this.#formatWmsFeatureInfoResult(featureMember, clickCoordinate);
           return featureInfoResult;
         }
       }
@@ -397,16 +397,11 @@ export class GVWMS extends AbstractGVRaster {
   /**
    * Translates the get feature information result set to the TypeFeatureInfoEntry[] used by GeoView.
    * @param {TypeJsonObject} featureMember - An object formatted using the query syntax.
-   * @param {OgcWmsLayerEntryConfig} layerConfig - The layer configuration.
    * @param {Coordinate} clickCoordinate - The coordinate where the user has clicked.
    * @returns {TypeFeatureInfoEntry[]} The feature info table.
    * @private
    */
-  #formatWmsFeatureInfoResult(
-    featureMember: TypeJsonObject,
-    layerConfig: OgcWmsLayerEntryConfig,
-    clickCoordinate: Coordinate
-  ): TypeFeatureInfoEntry[] {
+  #formatWmsFeatureInfoResult(featureMember: TypeJsonObject, clickCoordinate: Coordinate): TypeFeatureInfoEntry[] {
     const queryResult: TypeFeatureInfoEntry[] = [];
 
     let featureKeyCounter = 0;
